+++ examples/mixed_instances/outputs.tf
@@ -1,0 +1,6 @@
+output "outputs" {
+  value = [
+    module.mixed_instances,
+  ]
+  description = "Output for various resources in this module"
+}